Fix serializer modules
